Fix removeSingleSlash option adding slashes
Turns out the option caused single slashes to be added to URLs that had
none because the `URL` constructor defaults pathname to `/` both on parse
and .toString(), so I had to prevent overwriting `urlString` to be able
to check if the passed URL's pathname is `/`.

Also added an entry for the option to `index.test-d.ts`.